Add files via upload
small Python script that was added to a python caller transformer in an ETL for processing address data pulled from Rhode Island because the address number (house number) field was not parsed as prefix/number/suffix/unit etc. and was instead supplied as a single text field. This script is not robust enough to handle data from other states because it was based on observed patterns in the supplied data. The script may need to be adapted to fit your needs.